+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
-`2021-06-11`
+`2021-06-XX`
+
+### What's new
+
+- **Added `size` prop to `Input` and `Select` components.** Supports small and large sizes.
 
 ### Fixes